Soundness and performance rework
This commit reworks mfio core to be void of forced unsoundness. Instead, mfio will default to a sound, albeit potentially slower, implementation that cannot lead to undefined behavior, while enabling unsound optimizations through `mfio_assume_linear_types` config switch.

As the name implies, the unsound optimizations assume type linearity, i.e. that objects will never be forgotten, and will always be dropped. Without this assumption, it is unsafe to perform asynchronous completion I/O directly to stack variables, because it is impossible to know how long these variables will be valid for.

Aside from that, packet was optimized for the most common use cases. The optimizations entail adding a hardcoded "simple" packet code path for when a packet is merely a byte buffer. If a packet is simple, then vtables are not involved at all in the process of outputing data to the packet (or reading from the packet). However, flexibility is still maintained through the use of vtables on more complex packet types.